Adding auto lock heading
On toggle, the robot will auto lock heading to the nearest reef face so that the front of the bot will be parallel with the closest face. As to not interfere with path planner if the angles of autolock and path planner differ by ~5 degrees then it will auto disable lock heading.